Allow users to register cache handlers with SimplePie_Cache.
Instead of overriding SimplePie_Cache, you can now register your handler
directly with it. This is done by registering a prefix (the part of the
location before a colon) with a class to be instantiated as the cache
object. This class must implement the SimplePie_Cache_Base interface,
and this will be inforced in a later version of the code.