Update IRI/IRITest classes from Requests
This fixes a whole lot of tests for the IRI classes in two ways:
firstly, it changes the incorrect tests. Secondly, it fixes some of the
internals of the IRI class to make them more correct.

Almost all tests should now pass. Full passing for the IRI class coming
soon.