Ensure ipath normalisation is correct
This fixes all remaining IRI tests. See #195.

For HTTP URIs, the path should always be normalised to / when forming a
URI. We do this here rather than in `scheme_normalisation()` as it makes
more sense from a comparison point of view.

We also make sure that the host is not empty (null or empty string).
This may be incorrect, but it matches the tests and it's not really a
valid usecase that we care about.