Double-escape <title> in RSS feeds
When we have something like &amp;amp; the XML parser will decode this to
&amp;, correctly. However, when we use a heuristic later in the
Sanitizer, this is detected as HTML and not escaped back. In order to
fix this, we simply treat it as text way back in the Parser.

This *may* cause bugs, and will probably break things. More checking
will be required.