Ensure IRI::get_iri() returns a valid URI
Technically, this is a bad idea. IRI::get_iri() should always return a
valid normalized IRI. By converting it to a URI, this causes it to be
non-normalized. However, since we're only ever using it in a URI
context, this is something I'm willing to live with.

See also 9c28bdc652a84a052c9e8fe72043408369dc8a68 for `master`. Unlike
there, we can't use IRI::get_uri() instead, as that would be a break in
backwards compatibility.

Also see http://core.trac.wordpress.org/ticket/20999 for discussion