[FreshRSS] MD5-based caching
For feeds that do not support HTTP conditional requests, makes a MD5
hash to detect changes.
To do that, some cleaning is needed, e.g. removing comments such as
`<!-- Generated on XX -->`, or `<lastBuildDate>` which may be changing
even when no actual content has changed.

https://github.com/FreshRSS/FreshRSS/commit/9aab83af115de3b210ea41caae49b70a0f82492a

Take also care of the caching of feeds with errors. Before the cache
system was not used for feeds with errors; there was no protection
against repetitive refresh.

https://github.com/FreshRSS/FreshRSS/commit/00127f07c5fc784130d658e3f26519b0279fc6b8

Used in production in FreshRSS