Return effective feed URL when asking for non-permanent subscribe_url
`subscribe_url` method supports two modes: permanent and non-permanent.
And while the URL for the first one is obtained from the `File` after
it has been fetched, discounting auto-discovery, the latter one remains
the same from the time `set_feed_url` was called.

Because even the permanent mode concerns itself with redirection (although
limited to a potential 301 as a first link in the chain), I would expect
the non-permanent one to care doubly so.

This patch enables just that – after the feed is fetched, url property
is updated, regardless of auto-discovery being done.