Distinct cache for force_feed
Ensure that feeds retrieved with the force_feed option do not share the
same cache than feeds retrieved without this option.
Otherwise, in a context with multiple clients, clients without the
force_feed option would have a behaviour (working or not) depending
entirely on what other clients with force_feed are doing. In other
words, it procudes a side-effect difficult to predict, which should be
avoided.