Fix permanent_url for HTTP 301 (#660)
Using cURL CURLOPT_FOLLOWLOCATION instead of the manual method in
SimplePie hides the list of HTTP redirects along the way, and prevents
the distinction of e.g. 301 vs. 302 redirects needed for
`simplepie->subscribe_url(true)`.

This patch disables CURLOPT_FOLLOWLOCATION in SimplePie, and fixes the
manual method at the same time (curl_options where not given to
following redirection requests).
The use of CURLOPT_FOLLOWLOCATION was nevertheless not systematic (only
when open_basedir was not set), so now there is only one logic
independent of open_basedir.