Tests: switch out @expectedException annotations
These annotations are no longer supported in recent PHPUnit versions.

For PHP native errors/warnings/notices, the `TestCase::expectError()`, `TestCase::expectWarning()`, `TestCase::expectNotice()` and `TestCase::expectDeprecation()` methods (and variants) should be used.

For (custom) exceptions, the `TestCase::expectException()` method (and variants) should be used.

This also removes the need for the `setUpBeforeClass()` method in the `IRITest`, which only created an alias for a PHPUnit class.

The new methods and their variants are polyfilled via the PHPUnit Polyfill repo via the included trait.