EncodingTest: rename a custom assertion
The method `assertEquals()` is a PHPUnit native method and if it is overloaded, the method signature is expected to be the same as the method signature in PHPUnit itself.

However, the method signature for this method has changed a couple of times across PHPUnit versions.

The simplest way to get round this is to rename the custom assertion to not overlap with a PHPUnit native method name.