CacheTest: handle different exceptions PHP cross-version
The legacy manner of caching was PHP deprecated for quite a while already, but can no longer be supported in PHP 8 as support for calling a non-static method statically has been removed.

This updates the inline comments in the `SimplePie_Registry::call()` method along those lines and adjusts the exception expected by the `CacheTest::testDirectOverrideLegacy()` to match.