GH Actions: don't allow builds to fail on PHP 8.0 or 8.1
Now the test suite has been made cross-version compatible AND the PHP 8.1 incompatibility has been fixed AND the PHP 8.0 feature removal causing a test failure has been dealt with... the tests will run and pass on PHP 8.0 and 8.1.

Let's keep it that way by not allowing them to fail anymore.