SimplePie_Cache_Redis: fix bug for using undefined property
The `$ttl` property is not declared, not set on this cache, nor its parent, which means that it would always be passed as `null`.

As the call to the `Redis::expire()` method is within a `if ($this->options['expire'])` condition and a similar call to the `Redis::expire()` method in the `SimplePie_Cache_Redis::save()` method, also uses `$this->options['expire']` for the "ttl" value, changing the passed parameter to `$this->options['expire']` seems the correct fix.