Make status_code nullable
This is to distinguish a status_code being set on the type level,
making it harder to forget the case – static analysis would notify you.

To make it more consistent, we also change the few cases that could make
`status_code` contain a `0` to replace it with `null`:

 - `CURLINFO_HTTP_CODE` can contain `0`, but only when the request fails
   with an error such as `CURLE_COULDNT_RESOLVE_HOST`, in which case
   the File constructor will fail.
 - HTTP parser, which could return `0` when a web server decides to
   return it as a status code for some reason.
   I added a code that will fail for `0` status code
   (though other invalid HTTP response codes may still appear successful).
 - Successful local requests were already using `null` correctly.

While at it, simplify few conditionals using de Morgan’s laws
for easier reasoning.

This is technically a BC break but so is https://github.com/simplepie/simplepie/pull/728
and `null` behaves just like `0` from the perspective of non-strict comparison
so I would not expect issues in legacy applications.