File: Implement `Response` interface
Implementing the PSR-7 inspired `Response` interface gets us one step
closer to pluggable HTTP implementation.

Most of the complexity comes from the fact that PSR-7 `MessageInterface`
expects values of each header to always be a list, while `File` stored
each as a single strings. But we cannot change the type of the `$headers`
property without breaking backwards compatibility (especially for `File`
subclasses that modify the headers), and we cannot just confidently split
the header lines in the methods, since the concatenating multiple instances
of the same header like `HTTP\Parser` currently does is a lossy operation.

As a result, we have to check if `$headers` property changed every time
we access the internal header state.

Co-authored-by: Jan Tojnar <jtojnar@gmail.com>