Set SimplePie->$file to private
We are going to allow using HTTP Clients that will set this to a different
`Response` subclasses than just `File`.

It was already marked private with PHPDoc, let’s enforce it.
Hopefully, this will produce a more scrutable error message
in case someone wants to access `File` properties on this.