Remove package phpdoc annotation
`@subpackage` is deprecated and `@package` is mostly redundant with PHP namespaces.

And while the annotations sometimes do not match the namespaces perfectly,
which could be taken as a separate logical structure,
having two separate hierarchies is IMO more confusing than helpful.
Especially when the annotations are not even picked up by ApiGen.

https://docs.phpdoc.org/2.9/references/phpdoc/tags/package.html
https://docs.phpdoc.org/2.9/references/phpdoc/tags/subpackage.html
https://github.com/ApiGen/ApiGen/issues/1081

Removed with the following commands:

    sd '(?m)(?:\n^[ \t]+\*)?\n[ \t]+\*[ \t]+@package.+?(?:\n[ \t]+\*[ \t]+@subpackage.+)?(\n[ \t]+\*/|(\n[ \t]*\*))' '$2$1' "$(rg -l @subpackage -t php)"