Replace the licence blurbs with SPDX comments
SPDX is a standard for describing licensing and copyright status of project files that has been gaining traction lately.
In addition to not filling more than half a screen, it is machine readable and can be used to produce Software Bill of Materials.
We are keeping the long header in the compiled single-file bundle since that might be copied into third-party projects in a way that makes it easy to accidentally forget also copying the license text.

https://reuse.software/tutorial/
https://www.linuxfoundation.org/blog/blog/solving-license-compliance-at-the-source-adding-spdx-license-ids