HTTP/Parser: Support PSR-7 header format
PSR-7 `MessageInterface` expects values of each header is always a list.
But concatenating multiple instances of the same header like we do now
is a lossy operation so we cannot confidently split them.

To allow us to use PSR-7-like semantics internally,
let’s add PSR-7 mode to the HTTP parser that will avoid the imploding.