SimplePie: Do not reduce timeout for cache verification
Dividing a number may return a float, which will cause a `TypeError` once we add type hints.

The value was `1` originally, when timeouts were introduced along with `SimplePie_File` in
https://github.com/simplepie/simplepie/commit/5bf1814b6943be7708cf7689f01c1c6aeb901afe

We could consider hardcoding it again but we still want to keep it proportional to the master timeout
to allow people with slow connection to increase it, as introduced in
https://github.com/simplepie/simplepie/commit/fe7e33301e0ac1df982b4652b4b2ae3372e2d254

In fact, reducing the timeout probably does not make sense in the first place.
Even if we assume majority of cached resources will not get outdated,
waiting for server to reply `HTTP 304 Not Modified` can still take time
(domain name resolution can be particularly slow).
And if it turns out that the cached version is outdated,
the web server will try to send an updated version.
Having too low timeout can mean the request will be terminated,
only to be started again with the original timeout.

The timeout reduction was presumably done to prevent
fetching a cached resource from an unresponsive server without `force_cache_fallback`
from essentially doubling the timeout due to re-running the failed request.

Let’s use the original timeout and prevent repeat requests instead.