HTTP: Add `Response::with_header()` method
This is similar to `Psr\Http\Message\MessageInterface::withHeader()`
but does not preserve the casing of the header.

It will be useful for storing metadata (e.g. WebSub link) in the response.

We cannot use `static` as return type hint since it is only supported in PHP 8.0:
https://php.watch/versions/8.0/static-return-type
Nor can we use `self`, as that requires return covariance introduced in PHP 7.4:
https://wiki.php.net/rfc/covariant-returns-and-contravariant-parameters