PHPStan: Ignore incompatible PHPUnit argument type on PHP ≤ 7.4
In psr/simple-cache < 2.0.0, which is installed on PHP 7.4 or lower,
the `Psr\SimpleCache\InvalidArgumentException` interface does not extend `Throwable`.
But the current versions of PHPUnit require subtypes of `Throwable` to be given to `expectException()` method.
As a result, PHPStan will complain when we try to expect the non-`Throwable`-inheriting interface:

    Parameter #1 $exception of method PHPUnit\Framework\TestCase::expectException() expects class-string<Throwable>, string given.