ci: Add PHP 8.4
PHP 8.4 is in beta, with final version scheduled for November so it is time to start testing it.

Add it to CI matrix, allowing failures for now. Do not allow failures for PHP 8.3 since that has been stable for ages.

(cherry picked from commit 6b97dec75ed45c11a7411616643f855fe13c2add)