PHPStan: Add extension for typing Registry::call()
`Registry::call()` is widely used in SimplePie to allow extensibility. Unfortunately, the pattern hides the actual method from PHPStan so it cannot determine the return value type, and there is no annotation we can add to make it work. , This results in an inadequate analysis.

Thankfully, most cases allow us to determine the method statically so we can create a PHPStan extension doing just that.

We use `SimplePieUtils` instead of `Utils` namespace since this extension might be useful for some SimplePie consumers and do not want to subject them to symbol conflicts. However, we do not currently want to commit to making this a proper Composer package so use it on your own risk, stability not guaranteed.

We may want to add more type checking extensions in the future.