FileClient: Merge HTTP headers from cURL options (#912)
Previously, if a library user passed `CURLOPT_HTTPHEADER` option through `SimplePie::set_curl_options()`, it would discard all headers passed by SimplePie through the `$headers` parameter. Some of those are essential so discarding them is a bad idea.

Let’s append the externally provided HTTP headers to the list to be passed to `CURLOPT_HTTPHEADER` instead.

There is no simple way to remove the SimplePie-generated headers now but, since it only sets `Accept`, `if-modified-since`, `if-none-match` and `X-FORWARDED-FOR`, it is probably not a big deal. Especially when `FileClient` is about to be deprecated.