HTTP: Remove redundant doc comments from Response implementations
As per https://stackoverflow.com/a/26554272 and https://docs.phpdoc.org/guide/guides/inheritance.html, phpDocumentor will inherit the doc comments from the interface. Keeping them in implementation is thus mostly pointless noise. The only benefit is that the docs are immediately available even when not using an IDE or LSP but keeping them in sync is too annoying to be worth it.